# swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/game/simple":{"get":{"operationId":"GameController_getAllSimpleEquationGame","summary":"Fetch all simple equation game.","description":"Fetch a game given its id.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}},"403":{"description":"Invalid API Key"}},"tags":["game"],"security":[{"api-key":[]}]},"post":{"operationId":"GameController_createSimpleEquationGame","summary":"Create a new simple equation game.","description":"Create a new simple equation game.","parameters":[],"responses":{"201":{"description":"Resource created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimpleEquationGameResponseDto"}}}},"400":{"description":"Invalid input"},"403":{"description":"Invalid API Key"}},"tags":["game"],"security":[{"api-key":[]}]}},"/api/v1/game/simple/{gameId}":{"get":{"operationId":"GameController_getSimpleEquationGame","summary":"Fetch a specific simple equation game given its id.","description":"Fetch a game given its id.","parameters":[{"name":"gameId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}},"403":{"description":"Invalid API Key"}},"tags":["game"],"security":[{"api-key":[]}]}}},"info":{"title":"Equati math games","description":"Equati game API","version":"1.0.1","contact":{"name":"Contact the developer","url":"","email":"info@equati.de"},"termsOfService":"http://swagger.io/terms/","license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"tags":[{"name":"game","description":"Endpoint to create new and fetch existing games."}],"servers":[{"url":"http://localhost:8080"},{"url":"https://api.equati.de"}],"components":{"securitySchemes":{"api-key":{"type":"apiKey","in":"header","name":"X-API-KEY"}},"schemas":{"SimpleEquationGameResponseDto":{"type":"object","properties":{"operations":{"type":"array","items":{"type":"string"}},"possibleSolution":{"type":"array","items":{"type":"string"}},"source":{"type":"array","items":{"type":"number"}},"target":{"type":"number"}},"required":["operations","possibleSolution","source","target"]}}},"externalDocs":{"description":"Find out more about Swagger","url":"http://swagger.io/"}}